Setup keychain for terminal/ssh and X logins
Ideally the call to keychain in ~/.shinit would be removed but I
wanted to be able to run my password manager and paste passwords
into the keychain prompt. This doesn't work with X versions of
ssh-askpass  so for now this hack works.

One fix might be to spawn a terminal to add the keys in ~/.xsession
but I don't think we can avoid the call to keychain in ~/.shinit
because we need to make sure all X terminals have a chance to load
the keys.

Bottom line is that this still needs work.