Change how and when keychain is launched
For a console login start agents and load keys in `.profile`

For a display manager login start the agents in `.xsession` and
load the keys in `.xinitrc`. `.xinitrc` only loads keys if
`startx` was not run.

Doing things this way means no shell should have to run `keychain`
to attach to agents because all the necessary variables should be
inherited from either `.profile` or `.xsession`.

Going through this I also removed the `_xlogin` variable that was
set in `.xsession` and replaced it with a simple test to see if
`startx` had been run.